Re-organizing Auth Unit Tests (#149)
* Re-organizing auth unit tests

* Moving _request to TokenVerifier class

* Updated tests
